+++ app/src/main/java/com/whmnrc/qiangbizhong/model/bean/KillGoodsBean.java
@@ -70,13 +70,13 @@ public class KillGoodsBean {
         private Object Goods_Parameter;
         private String GoodsPrice_SpecName;
         private String GoodsPrice_AttrName;
-        private int GoodsPrice_VirtualPrice;
+        private double GoodsPrice_VirtualPrice;
         private String RushStartTime;
         private String RushEndTime;
         private int RushNumber;
         private int Bond;
         private int Price;
-        private int GoodsPrice_Price;
+        private double GoodsPrice_Price;
         private int IsEnd;
 
         public String getRushId() {
@@ -175,11 +175,11 @@ public class KillGoodsBean {
             this.GoodsPrice_AttrName = GoodsPrice_AttrName;
         }
 
-        public int getGoodsPrice_VirtualPrice() {
+        public double getGoodsPrice_VirtualPrice() {
             return GoodsPrice_VirtualPrice;
         }
 
-        public void setGoodsPrice_VirtualPrice(int GoodsPrice_VirtualPrice) {
+        public void setGoodsPrice_VirtualPrice(double GoodsPrice_VirtualPrice) {
             this.GoodsPrice_VirtualPrice = GoodsPrice_VirtualPrice;
         }
 
@@ -223,11 +223,11 @@ public class KillGoodsBean {
             this.Price = Price;
         }
 
-        public int getGoodsPrice_Price() {
+        public double getGoodsPrice_Price() {
             return GoodsPrice_Price;
         }
 
-        public void setGoodsPrice_Price(int GoodsPrice_Price) {
+        public void setGoodsPrice_Price(double GoodsPrice_Price) {
             this.GoodsPrice_Price = GoodsPrice_Price;
         }
 
@@ -276,13 +276,13 @@ public class KillGoodsBean {
         private Object Goods_Parameter;
         private String GoodsPrice_SpecName;
         private String GoodsPrice_AttrName;
-        private int GoodsPrice_VirtualPrice;
+        private double GoodsPrice_VirtualPrice;
         private String RushStartTime;
         private String RushEndTime;
         private int RushNumber;
         private int Bond;
         private int Price;
-        private int GoodsPrice_Price;
+        private double GoodsPrice_Price;
         private int IsEnd;
 
         public String getRushId() {
@@ -381,11 +381,11 @@ public class KillGoodsBean {
             this.GoodsPrice_AttrName = GoodsPrice_AttrName;
         }
 
-        public int getGoodsPrice_VirtualPrice() {
+        public double getGoodsPrice_VirtualPrice() {
             return GoodsPrice_VirtualPrice;
         }
 
-        public void setGoodsPrice_VirtualPrice(int GoodsPrice_VirtualPrice) {
+        public void setGoodsPrice_VirtualPrice(double GoodsPrice_VirtualPrice) {
             this.GoodsPrice_VirtualPrice = GoodsPrice_VirtualPrice;
         }
 
@@ -429,11 +429,11 @@ public class KillGoodsBean {
             this.Price = Price;
         }
 
-        public int getGoodsPrice_Price() {
+        public double getGoodsPrice_Price() {
             return GoodsPrice_Price;
         }
 
-        public void setGoodsPrice_Price(int GoodsPrice_Price) {
+        public void setGoodsPrice_Price(double GoodsPrice_Price) {
             this.GoodsPrice_Price = GoodsPrice_Price;
         }
