+++ app/src/main/java/com/whmnrc/qiangbizhong/model/bean/LuckDrawBean.java
@@ -126,8 +126,8 @@ public class LuckDrawBean {
         private Object Goods_Parameter;
         private String GoodsPrice_SpecName;
         private String GoodsPrice_AttrName;
-        private int GoodsPrice_Price;
-        private int GoodsPrice_VirtualPrice;
+        private double GoodsPrice_Price;
+        private double GoodsPrice_VirtualPrice;
         private int GoodsPrice_Stock;
         private int AwardPeopleCount;
         private int NeedCount;
@@ -215,19 +215,19 @@ public class LuckDrawBean {
             this.GoodsPrice_AttrName = GoodsPrice_AttrName;
         }
 
-        public int getGoodsPrice_Price() {
+        public double getGoodsPrice_Price() {
             return GoodsPrice_Price;
         }
 
-        public void setGoodsPrice_Price(int GoodsPrice_Price) {
+        public void setGoodsPrice_Price(double GoodsPrice_Price) {
             this.GoodsPrice_Price = GoodsPrice_Price;
         }
 
-        public int getGoodsPrice_VirtualPrice() {
+        public double getGoodsPrice_VirtualPrice() {
             return GoodsPrice_VirtualPrice;
         }
 
-        public void setGoodsPrice_VirtualPrice(int GoodsPrice_VirtualPrice) {
+        public void setGoodsPrice_VirtualPrice(double GoodsPrice_VirtualPrice) {
             this.GoodsPrice_VirtualPrice = GoodsPrice_VirtualPrice;
         }
