+++ app/src/main/java/com/whmnrc/qiangbizhong/model/bean/RushRecordBean.java
@@ -42,8 +42,8 @@ public class RushRecordBean {
     private int IsEnd;
     private int RushNumber;
     private int Order_Money;
-    private int GoodsPrice_Price;
-    private int GoodsPrice_VirtualPrice;
+    private double GoodsPrice_Price;
+    private double GoodsPrice_VirtualPrice;
     private String RushEndTime;
     private String RushStartTime;
     private String Goods_ImaPath;
@@ -152,19 +152,19 @@ public class RushRecordBean {
         this.Order_Money = Order_Money;
     }
 
-    public int getGoodsPrice_Price() {
+    public double getGoodsPrice_Price() {
         return GoodsPrice_Price;
     }
 
-    public void setGoodsPrice_Price(int GoodsPrice_Price) {
+    public void setGoodsPrice_Price(double GoodsPrice_Price) {
         this.GoodsPrice_Price = GoodsPrice_Price;
     }
 
-    public int getGoodsPrice_VirtualPrice() {
+    public double getGoodsPrice_VirtualPrice() {
         return GoodsPrice_VirtualPrice;
     }
 
-    public void setGoodsPrice_VirtualPrice(int GoodsPrice_VirtualPrice) {
+    public void setGoodsPrice_VirtualPrice(double GoodsPrice_VirtualPrice) {
         this.GoodsPrice_VirtualPrice = GoodsPrice_VirtualPrice;
     }
