+++ app/src/main/java/com/whmnrc/qiangbizhong/model/bean/ShopDetailsBean.java
@@ -203,7 +203,7 @@ public class ShopDetailsBean {
          */
 
         private int Goods_ShopType;
-        private int Goods_PriceMax;
+        private double Goods_PriceMax;
         private String Goods_ID;
         private String Goods_Content;
         private boolean Goods_IsOn;
@@ -214,7 +214,7 @@ public class ShopDetailsBean {
         private String Goods_ImaPath;
         private String StoreId;
         private int Goods_MonthCount;
-        private int Goods_PriceMin;
+        private double Goods_PriceMin;
         private boolean Goods_IsBuy;
         private int Goods_Sort;
         private int Goods_LimitCount;
@@ -229,11 +229,11 @@ public class ShopDetailsBean {
             this.Goods_ShopType = Goods_ShopType;
         }
 
-        public int getGoods_PriceMax() {
+        public double getGoods_PriceMax() {
             return Goods_PriceMax;
         }
 
-        public void setGoods_PriceMax(int Goods_PriceMax) {
+        public void setGoods_PriceMax(double Goods_PriceMax) {
             this.Goods_PriceMax = Goods_PriceMax;
         }
 
@@ -317,11 +317,11 @@ public class ShopDetailsBean {
             this.Goods_MonthCount = Goods_MonthCount;
         }
 
-        public int getGoods_PriceMin() {
+        public double getGoods_PriceMin() {
             return Goods_PriceMin;
         }
 
-        public void setGoods_PriceMin(int Goods_PriceMin) {
+        public void setGoods_PriceMin(double Goods_PriceMin) {
             this.Goods_PriceMin = Goods_PriceMin;
         }
