+++ app/src/main/java/com/whmnrc/qiangbizhong/model/bean/YiMeiIndexBean.java
@@ -179,10 +179,10 @@ public class YiMeiIndexBean {
             private boolean Goods_IsOn;
             private boolean Goods_IsBuy;
             private int GoodsPrice_Stock;
-            private int GoodsPrice_Price;
+            private double GoodsPrice_Price;
             private String GoodsPrice_SpecName;
             private String GoodsPrice_AttrName;
-            private int GoodsPrice_VirtualPrice;
+            private double GoodsPrice_VirtualPrice;
 
             public String getId() {
                 return Id;
@@ -312,11 +312,11 @@ public class YiMeiIndexBean {
                 this.GoodsPrice_Stock = GoodsPrice_Stock;
             }
 
-            public int getGoodsPrice_Price() {
+            public double getGoodsPrice_Price() {
                 return GoodsPrice_Price;
             }
 
-            public void setGoodsPrice_Price(int GoodsPrice_Price) {
+            public void setGoodsPrice_Price(double GoodsPrice_Price) {
                 this.GoodsPrice_Price = GoodsPrice_Price;
             }
 
@@ -336,11 +336,11 @@ public class YiMeiIndexBean {
                 this.GoodsPrice_AttrName = GoodsPrice_AttrName;
             }
 
-            public int getGoodsPrice_VirtualPrice() {
+            public double getGoodsPrice_VirtualPrice() {
                 return GoodsPrice_VirtualPrice;
             }
 
-            public void setGoodsPrice_VirtualPrice(int GoodsPrice_VirtualPrice) {
+            public void setGoodsPrice_VirtualPrice(double GoodsPrice_VirtualPrice) {
                 this.GoodsPrice_VirtualPrice = GoodsPrice_VirtualPrice;
             }
         }
